+++ src/components/form/form.config.js
@@ -1,4 +1,3 @@
-
 export const inputs = {
   email: {
     element: 'input',
@@ -9,9 +8,9 @@ export const inputs = {
       name: 'email',
       type: 'email',
       placeholder: 'email@address.com',
-      required: true
+      required: true,
     },
-    error: 'Use a valid email address (e.g., your_email@xyz.com).'
+    error: 'Use a valid email address (e.g., your_email@xyz.com).',
   },
   country: {
     element: 'select',
@@ -20,19 +19,19 @@ export const inputs = {
       id: 'country',
       class: 'form_input',
       name: 'country',
-      required: true
+      required: true,
     },
     children: {
       element: 'option',
       countries: [
-        { '': ' '},
+        { '': ' ' },
         { ch: 'Switzerland' },
-        { gb: 'United Kingdom'},
-        { us: 'United States'},
-        { ca: 'Canada'},
-      ]
+        { gb: 'United Kingdom' },
+        { us: 'United States' },
+        { ca: 'Canada' },
+      ],
     },
-    error: 'Select a country.'
+    error: 'Select a country.',
   },
   zipcode: {
     element: 'input',
@@ -42,9 +41,9 @@ export const inputs = {
       class: 'form_input',
       name: 'zipcode',
       type: 'text',
-      required: true
+      required: true,
     },
-    error: 'Enter a valid zip code corresponding to the selected country.'
+    // error: 'Select a country and enter a valid zip code corresponding to the selected country.',
   },
   password: {
     element: 'input',
@@ -56,9 +55,10 @@ export const inputs = {
       type: 'password',
       placeholder: 'Enter password',
       required: true,
-      pattern: '(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*\\W)(?!.* ).{8,30}'
+      pattern: '(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*\\W)(?!.* ).{8,30}',
     },
-    error: 'Password must be at least 8 characters long, can be up to 30 characters long, and must contain 1 of each of the following: uppercase letter, lowercase letter, number, and special character. No spaces allowed.'
+    error:
+      'Password must be at least 8 characters long, can be up to 30 characters long, and must contain 1 of each of the following: uppercase letter, lowercase letter, number, and special character. No spaces allowed.',
   },
   passwordConfirmation: {
     element: 'input',
@@ -69,23 +69,25 @@ export const inputs = {
       name: 'password_confirm',
       type: 'password',
       placeholder: 'Reenter password',
-      required: true
+      required: true,
     },
-    error: 'Passwords do not match.'
-  }
-}
+    error: 'Passwords do not match.',
+  },
+};
 
 export const formButtons = {
   cancel: {
     element: 'button',
     attributes: {
       class: 'btn_cancel',
-    }
+      type: 'button',
+    },
   },
   submit: {
     element: 'button',
     attributes: {
       class: 'btn_submit',
-    }
-  }
-}
+      type: 'submit',
+    },
+  },
+};