+++ src/components/form/form_validation_controller.js
@@ -2,32 +2,47 @@ import zipcodes from './zipcodes';
 
 export default () => ({
   resetForm() {
-    this.form.reset();
+    this.inputs.forEach((input) => {
+      input.value = '';
+      input.classList.remove('input_error');
+    });
+
+    this.validityErrors.forEach((error) => {
+      error.style.display = 'none';
+    });
   },
   submitForm(e) {
     e.preventDefault();
     this.inputs.forEach((input) => this.validateInput(input));
-
-    this.resetForm();
+    if ([...this.inputs].every((input) => this.validateInput(input))) {
+      console.log('form can be submitted');
+    } else {
+      console.log('form CANNOT be submitted');
+    }
+    // this.resetForm();
   },
   validateInput(e) {
     const input = e.currentTarget ? e.currentTarget : e;
+    const error = [...this.validityErrors].find((validityError) =>
+      validityError.classList.contains(input.id),
+    );
+
     if (input.id === 'password_confirm') {
       input.pattern = [...this.inputs].find((key) => key.type === input.type).value;
     } else if (input.id === 'zipcode') {
       const inputCountry = [...this.inputs].find((key) => key.id === 'country');
-      const iso = Object.keys(zipcodes).find((value) => value === inputCountry.value);
-      input.pattern = zipcodes[iso][0];
+      const zipCountry = Object.values(zipcodes).find(
+        (country) => country.iso === inputCountry.value,
+      );
 
-      // [...this.validityErrors].find(key => key.classList.contains(input.id)).textContent = 'test'
+      if (zipCountry) {
+        input.pattern = zipCountry.regex;
+        error.textContent = zipCountry.error;
+      }
     }
 
-    const error = [...this.validityErrors].find((validityError) =>
-      validityError.classList.contains(input.id),
-    );
     const validity = input.checkValidity();
     const { type } = e;
-    console.log(validity);
 
     if (!validity) {
       if (type === 'blur' && !input.classList.contains('input_error')) {
@@ -48,5 +63,7 @@ export default () => ({
       input.classList.remove('input_error');
       input.removeEventListener('input', this.validateInput);
     }
+
+    return validity;
   },
 });