+++ src/utilities/createElement.js
@@ -4,23 +4,23 @@ const setElementState = () => ({
   setAttributes(attributes) {
     Object.entries(attributes).forEach(([key, value]) => {
       this.setAttribute(key, value);
-    })
+    });
   },
-})
+});
 
 const BuildElement = (element) => {
   const state = {
     element,
-  }
+  };
 
   return {
-    ...setElementState(state)
-  }
-}
+    ...setElementState(state),
+  };
+};
 
 export default function createElement(element) {
   const htmlElement = document.createElement(element);
   Object.assign(htmlElement, BuildElement(htmlElement));
 
-  return htmlElement
+  return htmlElement;
 }